+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/model/PlaylistTrackEditModel.java
@@ -34,7 +34,7 @@ public final class PlaylistTrackEditModel extends SwingEntityEditModel {
 		super(PlaylistTrack.TYPE, connectionProvider);
 		persist(PlaylistTrack.TRACK_FK).set(false);
 		// Filter out tracks already in the current playlist
-		valueEvent(PlaylistTrack.PLAYLIST_FK).addConsumer(this::filterPlaylistTracks);
+		value(PlaylistTrack.PLAYLIST_FK).addConsumer(this::filterPlaylistTracks);
 	}
 
 	private void filterPlaylistTracks(Entity playlist) {